polymorphic method `register_model`
