+++ src/middlewared/middlewared/api/base/handler/model_provider.py
@@ -1,14 +1,18 @@
 from abc import ABC, abstractmethod
 import asyncio
 from concurrent.futures import Executor
+import functools
 import importlib
 import logging
 from types import ModuleType
-from typing import Callable
+from typing import Any, Callable, TypeAlias
 
 from middlewared.api.base import BaseModel
 
 logger = logging.getLogger(__name__)
+
+
+ModelFactory: TypeAlias = Callable[[type[BaseModel]], type[BaseModel]]
 
 
 class ModelProvider(ABC):
@@ -16,6 +20,14 @@ class ModelProvider(ABC):
     @abstractmethod
     def __init__(self):
         self.models: dict[str, type[BaseModel]]
+
+    def register_model(self, model_cls: type[BaseModel], *extra: Any) -> None:
+        """Register an API model.
+        
+        :param model_cls: The model class to register.
+        :param *extra: Extra arguments are ignored.
+        """
+        self.models[model_cls.__name__] = model_cls
 
     async def get_model(self, name: str) -> type[BaseModel]:
         """Get API model by name."""
@@ -49,7 +61,27 @@ class LazyModuleModelProvider(ModelProvider):
         self.models = None
         self.models_factories: dict[str, Callable[[], type[BaseModel]]] = {}
 
+    def register_model(self, model_cls: type[BaseModel], model_factory: ModelFactory, arg_model_name: str) -> None:
+        """Register a model factory to be called on `get_model`.
+
+        :param model_cls: The `BaseModel` class whose name should be used to register the model returned by
+            `model_factory`.
+        :param model_factory: A callable that returns the model.
+        :param arg_model_name: Name of the model to pass to `model_factory`.
+        """
+        self.models_factories[model_cls.__name__] = functools.partial(
+            _create_model, self, model_factory, arg_model_name
+        )
+
     async def get_model(self, name: str) -> type[BaseModel]:
+        """Retrieve an API model by its name.
+
+        Load the models from the module that this `ModelProvider` is responsible for. Either return the model from that
+        module or call the model factory that was registered with `name`.
+
+        :param name: Name of the model or name that its model factory was registered with.
+        :return: Either the model belonging to module `self.module_name` or the model returned by a model factory.
+        """
         async with self.lock:
             if self.models is None:
                 logger.debug(f"Importing {self.module_name!r}")
@@ -81,3 +113,16 @@ def models_from_module(module: ModuleType) -> dict[str, type[BaseModel]]:
         ]
         if isinstance(model, type) and issubclass(model, BaseModel)
     }
+
+
+def _create_model(model_provider: ModelProvider, model_factory: ModelFactory, arg_model_name: str) -> type[BaseModel]:
+    """Call a model factory.
+
+    :param model_provider: An instance of `ModelProvider` that contains the model to pass to `model_factory`.
+    :param model_factory: A callable that returns the model.
+    :param arg_model_name: Name of the model to pass to `model_factory`.
+
+    :raises KeyError: `arg_model_name` is not registered with `model_provider`.
+    """
+    arg = model_provider.models[arg_model_name]
+    return model_factory(arg)
